+++ packages/@tailwindcss-upgrade/src/index.ts
@@ -6,12 +6,13 @@ import path from 'node:path'
 import postcss from 'postcss'
 import { analyze as analyzeStylesheets } from './codemods/css/analyze'
 import { formatNodes } from './codemods/css/format-nodes'
+import { linkConfigs as linkConfigsToStylesheets } from './codemods/css/link'
 import { migrate as migrateStylesheet } from './codemods/css/migrate'
 import { sortBuckets } from './codemods/css/sort-buckets'
 import { migrate as migrateTemplate } from './codemods/template/migrate'
 import { prepareConfig } from './codemods/template/prepare-config'
 import { help } from './commands/help'
-import { linkConfigs as linkConfigsToStylesheets, split as splitStylesheets } from './migrate'
+import { split as splitStylesheets } from './migrate'
 import { migrateJsConfig } from './migrate-js-config'
 import { migratePostCSSConfig } from './migrate-postcss'
 import { migratePrettierPlugin } from './migrate-prettier'
